+++ src/main/kotlin/uk/co/ben_gibson/git/link/settings/ApplicationSettings.kt
@@ -48,6 +48,10 @@ class ApplicationSettings : PersistentStateComponent<ApplicationSettings?> {
         listeners.add(listener)
     }
 
+    fun incrementHits() {
+        hits++
+    }
+
     private fun notifyListeners() {
         listeners.forEach(ChangeListener::onChange)
     }